requirements.txt
biopython
requests
typer
tensorflow
dm-sonnet
graphs
